Add follow block endpoints

+++ routes/api.php
@@ -38,41 +38,43 @@ Route::middleware('api')->name('core.api.v1.')->prefix('v1')->group(function () 
         Route::match(['put', 'patch'], '/{media}', [MediaController::class, 'update'])->name('update');
         Route::match(['head', 'get'], '/{media}', [MediaController::class, 'show'])
             ->middleware('cache.headers:public;max_age=2628000;etag')->name('show');
-
-        Route::put('/{media}/like', [MediaLikeController::class, 'store'])->name('like');
-        Route::delete('/{media}/like', [MediaLikeController::class, 'destroy'])->name('unlike');
-
         Route::post('/{media}/report', [ReportController::class, 'media'])->name('report');
+        Route::middleware('auth:api')->group(function () {
+            Route::middleware('auth:api')->put('/{media}/like', [MediaLikeController::class, 'store'])->name('like');
+            Route::middleware('auth:api')->delete('/{media}/like', [MediaLikeController::class, 'destroy'])->name('unlike');
+        });
     });
 
     Route::prefix('user')->name('user.')->group(function () {
         Route::get('/{type}', [UserController::class, 'index'])
             ->where(['type' => '(likes|blocks|followers|followings|hashtags)'])->name('list');
 
-        Route::put('/{user}/block', [BlockController::class, 'store'])->name('block');
-        Route::delete('/{user}/block', [BlockController::class, 'destroy'])->name('unblock');
-
-        Route::put('/{user}/follow', [FollowController::class, 'store'])->name('follow');
-        Route::delete('/{user}/follow', [FollowController::class, 'destroy'])->name('unfollow');
-
         Route::post('/{user}/report', [ReportController::class, 'user'])->name('report');
 
         Route::get('/{user}/media', [MediaController::class, 'listByUser'])->name('media_list');
         Route::get('/{user}', [UserController::class, 'show'])->name('show');
+
+        Route::middleware('auth:api')->group(function () {
+            Route::put('/{user}/block', [BlockController::class, 'store'])->name('block');
+            Route::delete('/{user}/block', [BlockController::class, 'destroy'])->name('unblock');
+
+            Route::put('/{user}/follow', [FollowController::class, 'store'])->name('follow');
+            Route::delete('/{user}/follow', [FollowController::class, 'destroy'])->name('unfollow');
+        });
     });
 
-    Route::prefix('me')->name('profie.')->group(function () {
+    Route::middleware('auth:api')->prefix('me')->name('profie.')->group(function () {
         Route::get('/', [UserController::class, 'me']);
         Route::match(['put', 'patch'], '/', [UserController::class, 'update']);
         Route::delete('/', [UserController::class, 'destroy']);
         Route::get('/token', [UserController::class, 'token']);
     });
 
-    Route::match(['put', 'patch'], '/{alert}', [AlertController::class, 'update'])->name('alert.update');
+    Route::middleware('auth:api')->match(['put', 'patch'], '/{alert}', [AlertController::class, 'update'])->name('alert.update');
     Route::match(['put', 'patch'], '/change-password', [UserController::class, 'changePassword'])->name('user.change-password');
     Route::patch('/validate-otp', [UserController::class, 'validateOtp'])->name('user.validateOtp');
     Route::post('/request-otp', [UserController::class, 'requestOtp'])->name('user.requestOtp');
-    Route::delete('/logout', [UserController::class, 'logout'])->name('user.logout');
+    Route::middleware('auth:api')->delete('/logout', [UserController::class, 'logout'])->name('user.logout');
 
     Route::post('/login', [AuthController::class, 'login'])->name('user.login');
     Route::post('/register', [AuthController::class, 'register'])->name('user.register');
